requirements.txt
# Web Framework
streamlit>=1.28.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
kaleido>=0.2.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Testing
hypothesis>=6.90.0
pytest>=7.4.0

# Environment Variables
python-dotenv>=1.0.0

# HTTP Requests (for API integration)
requests>=2.31.0

# Alpaca API
alpaca-py>=0.20.0

# Sentiment Analysis
transformers>=4.30.0
torch>=2.0.0
